requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
python-dotenv>=0.19.0
click>=8.0.0
xlrd>=2.0.1
